separate customer and data models to their own files so they can be reused in other files
