+++ lib/models/verify_response_model.dart
@@ -1,5 +1,8 @@
 import 'dart:convert';
 
+import 'customer_model.dart';
+import 'data_model.dart';
+
 VerifyResponse verifyResponseFromJson(String str) =>
     VerifyResponse.fromJson(json.decode(str));
 
@@ -34,65 +37,5 @@ class VerifyResponse {
         "data": data!.toJson(),
         "message": message,
         "reference": reference,
-      };
-}
-
-class Customer {
-  Customer({
-    this.customerName,
-    this.email,
-    this.id,
-  });
-
-  String? customerName;
-  String? email;
-  int? id;
-
-  factory Customer.fromJson(Map<String, dynamic> json) => Customer(
-        customerName: json["customer_name"],
-        email: json["email"],
-        id: json["id"],
-      );
-
-  Map<String, dynamic> toJson() => {
-        "customer_name": customerName,
-        "email": email,
-        "id": id,
-      };
-}
-
-class Data {
-  Data({
-    this.amount,
-    this.channel,
-    this.currency,
-    this.numberUsed,
-    this.status,
-    this.transactionDate,
-  });
-
-  int? amount;
-  String? channel;
-  String? currency;
-  String? numberUsed;
-  String? status;
-  String? transactionDate;
-
-  factory Data.fromJson(Map<String, dynamic> json) => Data(
-        amount: json["amount"],
-        channel: json["channel"],
-        currency: json["currency"],
-        numberUsed: json["number_used"],
-        status: json["status"],
-        transactionDate: json["transaction_date"],
-      );
-
-  Map<String, dynamic> toJson() => {
-        "amount": amount,
-        "channel": channel,
-        "currency": currency,
-        "number_used": numberUsed,
-        "status": status,
-        "transaction_date": transactionDate,
       };
 }
